Extract floorplan render logic into editor 2D modules
