+++ packages/editor/src/lib/floorplan/items.ts
@@ -1,6 +1,11 @@
-import { getScaledDimensions } from '../schema'
-import type { AnyNode, AnyNodeId, ItemNode, LevelNode } from '../schema'
-import useLiveTransforms from '../store/use-live-transforms'
+import {
+  getScaledDimensions,
+  type AnyNode,
+  type AnyNodeId,
+  type ItemNode,
+  type LevelNode,
+  useLiveTransforms,
+} from '@pascal-app/core'
 import { getRotatedRectanglePolygon, rotatePlanVector } from './geometry'
 import type { FloorplanItemEntry, FloorplanNodeTransform, LevelDescendantMap } from './types'
