+++ packages/editor/src/lib/floorplan/walls.ts
@@ -1,4 +1,4 @@
-import type { WallNode } from '../schema'
+import type { WallNode } from '@pascal-app/core'
 
 const FLOORPLAN_WALL_THICKNESS_SCALE = 1.18
 const FLOORPLAN_MIN_VISIBLE_WALL_THICKNESS = 0.13
